refactor: rename SyncProgressState file to sync_progress_state
